# Conflicts:
#	CMakeLists.txt

+++ renderers/agg/src/agg_font_freetype.cpp
@@ -422,12 +422,12 @@ namespace mapserver
             y += bitmap.rows;
             pitch = -pitch;
         }
-        for(i = 0; i < bitmap.rows; i++)
+        for(i = 0; i < (unsigned int)bitmap.rows; i++)
         {
             sl.reset_spans();
             bitset_iterator bits(buf, 0);
             unsigned int j;
-            for(j = 0; j < bitmap.width; j++)
+            for(j = 0; j < (unsigned int)bitmap.width; j++)
             {
                 if(bits.bit()) sl.add_cell(x + j, cover_full);
                 ++bits;
@@ -463,11 +463,11 @@ namespace mapserver
             y += bitmap.rows;
             pitch = -pitch;
         }
-        for(i = 0; i < bitmap.rows; i++)
+        for(i = 0; i < (unsigned int)bitmap.rows; i++)
         {
             sl.reset_spans();
             const int8u* p = buf;
-            for(j = 0; j < bitmap.width; j++)
+            for(j = 0; j < (unsigned int)bitmap.width; j++)
             {
                 if(*p) sl.add_cell(x + j, ras.apply_gamma(*p));
                 ++p;
